Web Inspector: Inspector should use the last sourceURL / sourceMappingURL directive https://bugs.webkit.org/show_bug.cgi?id=153072 <rdar://problem/24168312>
Patch by Joseph Pecoraro <pecoraro@apple.com> on 2016-01-13
Reviewed by Timothy Hatcher.

Source/JavaScriptCore:

* parser/Lexer.cpp:
(JSC::Lexer<T>::parseCommentDirective):
Just keep overwriting the member variable so we end up with
the last directive value.

Source/WebInspectorUI:

* UserInterface/Base/Utilities.js:
Prepend instead of append the sourceURL so the provided string
can include a sourceURL and override it. Such as manually
evaluating a script through the quick console with a sourceURL.

* UserInterface/Models/SourceMapResource.js:
Fix an incorrectly named property.

LayoutTests:

* inspector/console/console-api-expected.txt:
* inspector/console/console-table-expected.txt:
* platform/mac/inspector/model/remote-object-expected.txt:
Updated line number in output now that code evaluated through the frontend
gets the __WebInspectorInternal__ sourceURL prepended to the start instead
of appended to the end.

* inspector/debugger/sourceURLs-expected.txt:
* inspector/debugger/sourceURLs.html:
Update the test to find the second sourceURL, not the first.

Identifier: 171158@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@194981 268f45cc-cd09-0410-ab3c-d52691b4dbfc