Equivalence PropertyCondition needs to check the offset it uses to load the value from is not invalidOffset https://bugs.webkit.org/show_bug.cgi?id=152912
Reviewed by Mark Lam.

When checking the validity of an Equivalence PropertyCondition we do not check that the offset returned by
the structure of the object in the equivalence condition is valid. The offset might be wrong for many reasons.
The one we now test for is when the GlobalObject has a property that becomes a variable the property is deleted
thus the offset is now invalid.

* bytecode/PropertyCondition.cpp:
(JSC::PropertyCondition::isStillValidAssumingImpurePropertyWatchpoint):
* tests/stress/global-property-into-variable-get-from-scope.js: Added.


Identifier: 171428@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@195462 268f45cc-cd09-0410-ab3c-d52691b4dbfc