Handle soft spaces after accepted candidates https://bugs.webkit.org/show_bug.cgi?id=153331 -and corresponding- rdar://problem/23958418
Reviewed by Darin Adler.

Source/WebCore:

Candidates now come with built-in spaces, so we should not insert a space for
them.
* editing/Editor.cpp:
(WebCore::Editor::handleAcceptedCandidate):

New SPI to properly handle these soft spaces.
* platform/spi/mac/NSSpellCheckerSPI.h:

Source/WebKit2:

New member variable m_softSpaceRange keeps track to the NSRange of a soft
space if the last text that was inserted has a soft space at the end.
* UIProcess/Cocoa/WebViewImpl.h:
* UIProcess/Cocoa/WebViewImpl.mm:
(WebKit::WebViewImpl::WebViewImpl):

The space at the end of candidates is a soft space. If that space exists,
cache the range of the space in m_softSpaceRange.
(WebKit::WebViewImpl::handleAcceptedCandidate):

When new text is inserted, find out if it is being inserted right after a
soft space. If it is, then [NSSpellChecker deletesAutospaceBeforeString] will
tell us if the space needs to be removed. If that is the case, then set the
replacementString to the soft space.
(WebKit::WebViewImpl::insertText):



Identifier: 171506@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@195547 268f45cc-cd09-0410-ab3c-d52691b4dbfc