REGRESSION (r194557): Keyboard shortcuts stop working after the WKWebView is unparented and reparented https://bugs.webkit.org/show_bug.cgi?id=153492 <rdar://problem/24138989>
Reviewed by Dan Bernstein.

* UIProcess/ios/WKContentViewInteraction.h:
* UIProcess/ios/WKContentViewInteraction.mm:
(-[WKContentView canBecomeFirstResponder]):
(-[WKContentView becomeFirstResponder]):
(-[WKContentView resignFirstResponder]):
When WKWebView is unparented, WKContentView will attempt to resignFirstResponder upwards,
first asking WKWebView. After r194557, WKWebView will accept first responder and forward
it on to the WKContentView, which will happily accept it again, despite being the view
that's trying to resign. This will cause us to completely lose first responder,
where it was actually supposed to propagate up above WKWebView to the client.

Keep track of when WKContentView is resigning first responder, and don't
let it become first responder while it is doing so, breaking the cycle.


Identifier: 171556@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@195602 268f45cc-cd09-0410-ab3c-d52691b4dbfc